Remove 659 duplicate words from the dictionary
Dictionary is also alphabetized. They're chosen randomly, so it
shouldn't matter.

All it takes to perform this fix is to open vim and enter `:%sort u`.
This command takes the entire contents of the file and sorts it by line.
The `u` option removes duplicates.

(And to save and exit vim after running the command, you can just type
`ZZ`)